message 관련 파일들 정리
